Add ShoppingList and remove TODOs older
