+++ client/src/components/ContentModal/Bank/Bank.jsx
@@ -5,8 +5,6 @@ import { Loans, Overview, PayDebt, Save } from "./Forms";
 export default function ContentBank() {
   const [selected, setSelected] = React.useState("overview");
 
-  // TODO: In CardBody will be used FormComp
-  // TODO: Add bg into of each content.
   return (
     <Tabs aria-label="Options" selectedKey={selected} onSelectionChange={setSelected}>
       <Tab key="overview" title="Overview">
